Refactored nests.js.
